Extracted, generated (with stub) and persisted chosen CP endpoint
The chosen CloudProvider endpoint are generated after the choice has
been made, using the Provider and its Services from CMDB.
The CloudProviderEndpoint information is the passed through the WF in
the DeploymentMessage. It can be used by IM to generated authentication
data accordingly (NOT DONE YET).
The information is also persisted (serialized) in the Deployment for
other WFs (Update/Undeploy).

See #51